extract getCode to an outer file
